+++ src/builtins.h
@@ -27,6 +27,8 @@ Value b_range(VM *vm, int argc, Value *args);
 Value b_map(VM *vm, int argc, Value *args);
 Value b_filter(VM *vm, int argc, Value *args);
 Value b_reduce(VM *vm, int argc, Value *args);
+Value b_sql_query(VM *vm, int argc, Value *args);
+Value b_sql_write(VM *vm, int argc, Value *args);
 Value b_json(VM *vm, int argc, Value *args);
 Value b_stringify(VM *vm, int argc, Value *args);
 Value b_now(VM *vm, int argc, Value *args);