+++ src/interp.c
@@ -793,6 +793,8 @@ void bridge_seed_builtins(VM *vm) {
         {"map", b_map},
         {"filter", b_filter},
         {"reduce", b_reduce},
+        {"sql_query", b_sql_query},
+        {"sql_write", b_sql_write},
         {"json", b_json},
         {"stringify", b_stringify},
         {"now", b_now},